wip: refactor filter chips
