+++ src/ui/components/FilterChip/FilterChip.types.ts
@@ -1,0 +1,10 @@
+import { NotificationFilter } from "../../pages/Notifications/Notification.types";
+
+interface FilterChipProps {
+  filter: NotificationFilter;
+  label: string;
+  isActive: boolean;
+  onClick: (filter: NotificationFilter) => void;
+}
+
+export type { FilterChipProps };
