+++ src/ui/pages/Notifications/Notification.types.ts
@@ -12,13 +12,6 @@ interface NotificationItemProps {
   onOptionButtonClick: (item: KeriaNotification) => void;
 }
 
-interface FilterChipProps {
-  filter: NotificationFilter;
-  label: string;
-  isActive: boolean;
-  onClick: (filter: NotificationFilter) => void;
-}
-
 export { NotificationFilter };
 
-export type { NotificationItemProps, FilterChipProps };
+export type { NotificationItemProps };
